ready to deploy

+++ BackEnd/controllers/admin-controllers.js
@@ -24,31 +24,53 @@ const getAllAdmins = async (req, res, next) => {
   }
 };
 const getAdminById = async (req, res, next) => {
-  
   const id = req.params.id;
   const admin = await AdminModel.findById(id);
-  
-  if(admin){
-    res.json({ message: {message:"Ok",type:'success'}, data: admin });
-  }else{
-    res.json({
-      message:{message:'Id Is Not Correct',type:'failed'}
-    })
-  }
 
+  if (admin) {
+    res.json({ message: { message: "Ok", type: "success" }, data: admin });
+  } else {
+    res.json({
+      message: { message: "Id Is Not Correct", type: "failed" },
+    });
+  }
 };
 
 const deleteAdmin = async (req, res, next) => {
   if (req.body.key === process.env.API_KEY) {
     const admin = await AdminModel.findOne({ userName: req.body.userName });
+    const controllerAdmin = await AdminModel.findOne({
+      userName: req.body.controllerAdmin,
+    });
     if (admin) {
-      await AdminModel.deleteOne({ userName: req.body.userName });
-      res.json({
-        message: {
-          message: `${req.body.userName} Removed From Admin List`,
-          type: "success",
-        },
-      });
+      if (
+        (controllerAdmin.role === "Head Admin" && admin.role === "Admin") ||
+        controllerAdmin.role === "Senior Admin"
+      ) {
+       if(controllerAdmin.userName !== admin.userName){ await AdminModel.deleteOne({ userName: req.body.userName });
+        res.json({
+          message: {
+            message: `${req.body.userName} Removed From Admin List`,
+            type: "success",
+          },
+        });}else{
+           res.json({
+             message: {
+               message: "You Cant Remove Yourself XD",
+               type: "failed",
+             },
+             data: {},
+           });
+        }
+      } else {
+        res.json({
+          message: {
+            message: "You Cant Remove Admins With Same Or Higher Role",
+            type: "failed",
+          },
+          data: {},
+        });
+      }
     } else {
       res.json({ message: { message: "Not EXist", type: "failed" }, data: {} });
     }
@@ -62,14 +84,12 @@ const deleteAdmin = async (req, res, next) => {
   }
 };
 
-
-const createAdmin = async(req, res, next) => {
+const createAdmin = async (req, res, next) => {
   if (req.body.key === process.env.API_KEY) {
     if (await AdminModel.findOne({ userName: req.body.userName })) {
       res.json({
         message: {
-          message:
-            "Operation Failure This UserName Already Taken",
+          message: "Operation Failure This UserName Already Taken",
           type: "failed",
         },
       });
@@ -79,8 +99,8 @@ const createAdmin = async(req, res, next) => {
         role: req.body.role,
         userName: req.body.userName,
         password: req.body.password,
-        createdBy:req.body.createdBy,
-        creatorRole:req.body.creatorRole
+        createdBy: req.body.createdBy,
+        creatorRole: req.body.creatorRole,
       });
 
       await newAdmin.save();
